fix launchpad sell + indexer
