--- apps/indexer/src/constants.ts
@@ -1,8 +1,0 @@
-export const FACTORY_ADDRESS = '0x01a46467a9246f45c8c340f1f155266a26a71c07bd55d36e8d1c7d0d438a2dbc'
-export const STARTING_BLOCK = 615556
-
-
-
-export const LAUNCHPAD_ADDRESS = {
-    SEPOLIA:"0x74acb6752abb734a7b3388567429217988e02409d9bf43c5586dc2c4f8baf40",
-}
