--- apps/indexer/src/deps.ts
@@ -1,6 +1,0 @@
-export { ec, hash, uint256, shortString } from 'https://esm.sh/starknet@5.14'
-export { formatUnits } from 'https://esm.sh/viem@1.4'
-
-export type { Block, FieldElement, Filter } from 'https://esm.sh/@apibara/indexer@0.3/starknet'
-export type { Config, NetworkOptions } from "https://esm.sh/@apibara/indexer";
-export type { Console } from "https://esm.sh/@apibara/indexer/sink/console";
