+++ apps/indexer/src/script.js
