+++ onchain/src/types/launchpad_types.cairo
@@ -54,7 +54,8 @@ pub struct TokenLaunch {
     pub liquidity_raised: u256,
     pub token_holded: u256,
     pub is_liquidity_launch: bool,
-    pub slope:u256
+    pub slope: u256,
+    pub threshold_liquidity: u256,
 }
 
 #[derive(Drop, Serde, Copy, starknet::Store)]
